requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0

# Data & math
numpy>=1.24.0
pandas>=2.0.0

# Visualization & notebooks
matplotlib>=3.7.0
jupyter>=1.0.0
tqdm>=4.65.0

# Optimization for Qwen
bitsandbytes>=0.41.0
accelerate>=0.25.0

# For TrOCR
sentencepiece>=0.1.99
tiktoken>=0.5.0
protobuf>=7.0.0

# For metrics
scikit-learn>=1.3.0
python-Levenshtein>=0.20.0
psutil>=5.9.0